+++ ffquiz/common/src/main/java/com/ff/common/service/QuestionBankService.java
@@ -4,5 +4,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.ff.common.entity.po.QuestionBank;
 
 public interface QuestionBankService extends IService<QuestionBank> {
-    void rmById(Integer id) throws Exception;
+    void rmById(Long id) throws Exception;
 }
